cassis: split get_promoters into separate file and test file
